memperbaiki bug pada validasi

+++ database/migrations/2023_11_14_151743_create__nominals_topup.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('game_id');
             $table->string('nama_nominal');
-            $table->decimal('nominal', 10, 2);
+            $table->integer('nominal');
             $table->timestamps();
         });
     }
